Move LocalBlock to wal package
